Change to use hash for get and set glob

+++ src/data/ops.rs
@@ -14,8 +14,8 @@ use crate::utils::{err::*};
 pub enum Inst {
     OpReturn,
     OpConstant(usize), // idx in const pool, -> load idx onto stack
-    OpGetGlobal(usize),
-    OpSetGlobal(usize), // idx of variable name
+    OpGetGlobal(u64), // hash of variable name
+    OpSetGlobal(u64), // hash of variable name
     OpNegate,
     OpAdd,
     OpSub,
